fix(board): lift draft state to page + clear-each-on-save per Codex review (round 1)
(1) BoardView unmounts on a board↔list view switch, which destroyed
unsaved inline drafts and bypassed the leave guard. Move the draft state
(draftText/draftOpen) to the page and bind it into BoardView, and move
the beforeNavigate guard + Save/Discard/Stay dialog to the page (always
mounted). A draft now survives view switches (state persists; the card
just isn't rendered off-board) and the guard fires regardless of view.

(2) leaveSaveAll cleared all drafts only after the whole loop, so a
retry after a partial failure re-created already-saved drafts. Clear
each draft as its create succeeds.
