+++ web/src/routes/[username]/[workspace]/[collection]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { page } from '$app/state';
-	import { goto } from '$app/navigation';
+	import { goto, beforeNavigate } from '$app/navigation';
 	import { api, PadApiError, isPlanLimitError, planLimitMessage } from '$lib/api/client';
 	import type { BulkItemsRequest, Collection, Item, QuickAction, View, ViewConfig } from '$lib/types';
 	import { parseSettings, parseFields, parseSchema, parseTags, getStatusOptions, itemUrlId, formatItemRef } from '$lib/types';
@@ -1050,6 +1050,76 @@
 		}
 	}
 
+	// ── Inline draft cards: page-owned state + leave guard (TASK-1676) ──
+	// State lives here (not in BoardView) so a draft survives a
+	// board↔list view switch that unmounts BoardView, and so the leave
+	// guard + dialog stay mounted regardless of view. BoardView binds
+	// these and renders the per-lane draft cards.
+	let draftText = $state<Record<string, string>>({});
+	let draftOpen = $state<Record<string, boolean>>({});
+	let savingDrafts = $state(false);
+	let showLeaveDialog = $state(false);
+	let pendingNav = $state<(() => void) | null>(null);
+	let bypassNavGuard = false;
+
+	let hasUnsavedDrafts = $derived(Object.values(draftText).some((t) => t.trim().length > 0));
+
+	// Intercept in-app navigation while a draft is unsaved. `nav.to` is
+	// null for full unload (reload / tab close / external) — drafts are
+	// kept only until reload, so those aren't guarded.
+	beforeNavigate((nav) => {
+		if (bypassNavGuard || !hasUnsavedDrafts || !nav.to) return;
+		const url = nav.to.url;
+		nav.cancel();
+		pendingNav = () => {
+			bypassNavGuard = true;
+			goto(url).finally(() => { bypassNavGuard = false; });
+		};
+		showLeaveDialog = true;
+	});
+
+	function runPendingNav() {
+		const act = pendingNav;
+		pendingNav = null;
+		showLeaveDialog = false;
+		act?.();
+	}
+
+	async function leaveSaveAll() {
+		if (savingDrafts) return;
+		savingDrafts = true;
+		try {
+			// Clear EACH draft as its create succeeds (not all at the end),
+			// so retrying after a partial failure can't re-create the ones
+			// that already saved (Codex round 1).
+			for (const [col, text] of Object.entries(draftText)) {
+				const title = text.trim();
+				if (!title) continue;
+				await quickCreateInColumn(col, title, false);
+				delete draftText[col];
+				delete draftOpen[col];
+			}
+		} catch {
+			// A create failed (already toasted) — keep the dialog open with
+			// the still-unsaved drafts so the user can retry or discard.
+			savingDrafts = false;
+			return;
+		}
+		savingDrafts = false;
+		runPendingNav();
+	}
+
+	function leaveDiscard() {
+		draftText = {};
+		draftOpen = {};
+		runPendingNav();
+	}
+
+	function leaveStay() {
+		pendingNav = null;
+		showLeaveDialog = false;
+	}
+
 	async function quickCreate() {
 		const title = quickCreateTitle.trim();
 		if (!title || !wsSlug || !collSlug || creatingNew) return;
@@ -1963,6 +2033,8 @@
 				onGroupReorder={handleGroupReorder}
 				oncreate={canEditThisCollection ? openQuickCreate : undefined}
 				onCreateInColumn={canEditThisCollection ? quickCreateInColumn : undefined}
+				bind:draftText
+				bind:draftOpen
 				onMoveColumn={canBulkEdit ? handleBulkMove : undefined}
 				onTagColumn={canBulkEdit ? handleBulkTag : undefined}
 				onUntagColumn={canBulkEdit ? handleBulkUntag : undefined}
@@ -2041,7 +2113,88 @@
 	/>
 {/if}
 
+{#if showLeaveDialog}
+	<!-- Leave guard (TASK-1676): an in-app navigation was intercepted
+	     because a lane has an unsaved draft card. -->
+	<div
+		class="leave-overlay"
+		role="presentation"
+		onclick={(e) => { if (e.target === e.currentTarget) leaveStay(); }}
+		onkeydown={(e) => { if (e.key === 'Escape') leaveStay(); }}
+	>
+		<div class="leave-dialog" role="dialog" aria-modal="true" aria-label="Unsaved card" tabindex="-1">
+			<h3 class="leave-title">Unsaved card</h3>
+			<p class="leave-body">You have an unsaved card. Save it before leaving?</p>
+			<div class="leave-actions">
+				<button class="leave-stay" disabled={savingDrafts} onclick={leaveStay}>Stay</button>
+				<button class="leave-discard" disabled={savingDrafts} onclick={leaveDiscard}>Discard</button>
+				<button class="leave-save" disabled={savingDrafts} onclick={leaveSaveAll}>Save</button>
+			</div>
+		</div>
+	</div>
+{/if}
+
 <style>
+	.leave-overlay {
+		position: fixed;
+		inset: 0;
+		z-index: 1000;
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		background: rgba(0, 0, 0, 0.45);
+		padding: var(--space-4);
+	}
+	.leave-dialog {
+		width: 100%;
+		max-width: 360px;
+		padding: var(--space-4);
+		background: var(--bg-primary);
+		border: 1px solid var(--border);
+		border-radius: var(--radius-lg);
+		box-shadow: var(--shadow-lg, 0 10px 30px rgba(0, 0, 0, 0.3));
+	}
+	.leave-title {
+		margin: 0 0 var(--space-2);
+		font-size: 1em;
+		color: var(--text-primary);
+	}
+	.leave-body {
+		margin: 0 0 var(--space-4);
+		font-size: 0.875em;
+		color: var(--text-secondary);
+	}
+	.leave-actions {
+		display: flex;
+		justify-content: flex-end;
+		gap: var(--space-2);
+	}
+	.leave-actions button {
+		padding: 6px 14px;
+		border-radius: var(--radius-sm);
+		font-size: 0.8125em;
+		cursor: pointer;
+		border: 1px solid var(--border);
+	}
+	.leave-actions button:disabled {
+		opacity: 0.5;
+		cursor: default;
+	}
+	.leave-stay {
+		background: var(--bg-secondary);
+		color: var(--text-primary);
+	}
+	.leave-discard {
+		background: none;
+		border-color: var(--accent-red, #ef4444) !important;
+		color: var(--accent-red, #ef4444);
+	}
+	.leave-save {
+		background: var(--accent-blue);
+		border-color: var(--accent-blue) !important;
+		color: #fff;
+	}
+
 	.collection-page {
 		max-width: var(--content-max-width);
 		margin: 0 auto;
